LayoutTests/html5lib/runner.html

<!doctype html>
<script>
var test_files = [
  'resources/pending-spec-changes-plain-text-unsafe.dat',
  'resources/pending-spec-changes.dat',
  'resources/plain-text-unsafe.dat',
  'resources/tests_innerHTML_1.dat',
  'resources/webkit01.dat',
  'resources/webkit02.dat',
  'resources/doctype01.dat',
  'resources/scriptdata01.dat',
  'resources/html5test-com.dat',
  'resources/entities01.dat',
  'resources/entities02.dat',
  'resources/comments01.dat',
  'resources/adoption01.dat',
  'resources/adoption02.dat',
  'resources/inbody01.dat',
  'resources/isindex.dat',
  'resources/tables01.dat',
  'resources/tricky01.dat',
  'resources/scripted/adoption01.dat',
  'resources/scripted/webkit01.dat'
];
</script>
<script src=../resources/dump-as-markup.js></script>
<script src="resources/runner.js"></script>
